Add graphics and option to play again
Added images and CSS stylings to the buttons and page. Created 'play
again' button that will reset the scores when clicked.

The button will keep adding itself, so the next step will be to
remove event listeners when the button appears and to add them again
when button is clicked. This should make sure the button only appears
once.