Refractor code and get abortController working
"Play again" button now needs to be pressed in order for rock, paper
and scissors buttons to be clickable on winning a set of five matches.

Added css styling to buttons and text, and refractored code in
JavaScript file. Refractoring included making eventlisteners its
own function as it needed to be called when clicking the "Play Again"
button.

The code is still very clunky, particularly in the playMatch function,
but it works. playMatch function could be more efficient as
a loop, but it works.